LPS-54223 implement portlet monitoring control and pass into MonitoringPortlet
