+++ portal-service/src/com/liferay/portal/kernel/monitoring/DefaultPortletMonitoringControl.java
@@ -1,0 +1,76 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.monitoring;
+
+/**
+ * @author Michael C. Han
+ */
+public class DefaultPortletMonitoringControl
+	implements PortletMonitoringControl {
+
+	@Override
+	public boolean isMonitorPortletActionRequest() {
+		return _monitorPortletActionRequest;
+	}
+
+	@Override
+	public boolean isMonitorPortletEventRequest() {
+		return _monitorPortletEventRequest;
+	}
+
+	@Override
+	public boolean isMonitorPortletRenderRequest() {
+		return _monitorPortletRenderRequest;
+	}
+
+	@Override
+	public boolean isMonitorPortletResourceRequest() {
+		return _monitorPortletResourceRequest;
+	}
+
+	@Override
+	public void setMonitorPortletActionRequest(
+		boolean monitorPortletActionRequest) {
+
+		_monitorPortletActionRequest = monitorPortletActionRequest;
+	}
+
+	@Override
+	public void setMonitorPortletEventRequest(
+		boolean monitorPortletEventRequest) {
+
+		_monitorPortletEventRequest = monitorPortletEventRequest;
+	}
+
+	@Override
+	public void setMonitorPortletRenderRequest(
+		boolean monitorPortletRenderRequest) {
+
+		_monitorPortletRenderRequest = monitorPortletRenderRequest;
+	}
+
+	@Override
+	public void setMonitorPortletResourceRequest(
+		boolean monitorPortletResourceRequest) {
+
+		_monitorPortletResourceRequest = monitorPortletResourceRequest;
+	}
+
+	private boolean _monitorPortletActionRequest;
+	private boolean _monitorPortletEventRequest;
+	private boolean _monitorPortletRenderRequest;
+	private boolean _monitorPortletResourceRequest;
+
+}
